Split line-budget projection helpers
