+++ loopx/benchmark_ledger.py
@@ -3175,209 +3175,6 @@ def _iter_ledger_runs(ledger: dict[str, Any]) -> list[dict[str, Any]]:
     return runs
 
 
-def _ledger_score_value(run: dict[str, Any]) -> float | None:
-    value = run.get("official_score")
-    if isinstance(value, bool) or value is None:
-        return None
-    if isinstance(value, (int, float)):
-        return float(value)
-    if isinstance(value, str):
-        try:
-            return float(value)
-        except ValueError:
-            return None
-    return None
-
-
-def _current_aggregate_failure_label_list(run: dict[str, Any]) -> list[str]:
-    labels: list[str] = []
-    seen: set[str] = set()
-    for key in ("failure_labels", "failure_attribution_labels"):
-        for label in _compact_list(run.get(key), limit=16):
-            if label not in seen:
-                seen.add(label)
-                labels.append(label)
-    return labels
-
-
-def _current_aggregate_failure_labels(run: dict[str, Any]) -> set[str]:
-    return set(_current_aggregate_failure_label_list(run))
-
-
-def _current_aggregate_effective_failure_class(run: dict[str, Any]) -> str:
-    for key in (
-        "failure_class",
-        "score_failure_attribution",
-        "attempt_failure_class",
-        "attempt_failure_label",
-        "first_blocker",
-    ):
-        value = _compact_text(run.get(key), limit=120)
-        if value and value not in {"none", "None", "null"}:
-            return value
-    for label in _current_aggregate_failure_label_list(run):
-        if label and label not in {"none", "None", "null"}:
-            return label
-    return ""
-
-
-def _current_aggregate_bucket(run: dict[str, Any] | None) -> str:
-    if not isinstance(run, dict):
-        return "missing"
-    score = _ledger_score_value(run)
-    if score is not None:
-        if score >= 1.0:
-            return "pass"
-        if score > 0.0:
-            return "partial"
-        return "official_zero"
-    labels = _current_aggregate_failure_labels(run)
-    failure_class = _current_aggregate_effective_failure_class(run)
-    failure_scope = _compact_text(run.get("failure_scope"), limit=120)
-    score_status = _compact_text(run.get("score_status"), limit=120)
-    if (
-        "verifier_no_reward" in labels
-        or "verifier_reward_missing" in labels
-        or "reward_missing" in failure_class
-        or "verifier_no_reward" in failure_class
-    ):
-        return "verifier_no_reward"
-    if (
-        failure_scope == "runner_or_setup"
-        or failure_scope == "score_missing"
-        or "infrastructure" in failure_class
-        or "setup" in failure_class
-        or "preflight" in failure_class
-        or "compose" in failure_class
-        or "docker" in failure_class
-        or "runner" in failure_class
-        or any(
-            "infra" in label
-            or "setup" in label
-            or "compose" in label
-            or "docker" in label
-            for label in labels
-        )
-    ):
-        return "setup_runner_infra"
-    if score_status == "missing":
-        return "missing"
-    return "missing"
-
-
-_CURRENT_AGGREGATE_BUCKET_RANK = {
-    "missing": 0,
-    "setup_runner_infra": 1,
-    "verifier_no_reward": 2,
-    "official_zero": 3,
-    "partial": 4,
-    "pass": 5,
-}
-
-
-def _current_aggregate_run_sort_key(run: dict[str, Any]) -> tuple[Any, ...]:
-    bucket = _current_aggregate_bucket(run)
-    score = _ledger_score_value(run)
-    score_rank = score if score is not None else -1.0
-    return (
-        _CURRENT_AGGREGATE_BUCKET_RANK.get(bucket, 0),
-        score_rank,
-        _compact_text(run.get("recorded_at"), limit=80),
-        _compact_text(run.get("run_group_id"), limit=160),
-        _compact_text(run.get("run_id"), limit=80),
-    )
-
-
-def _current_aggregate_run_summary(run: dict[str, Any] | None) -> dict[str, Any]:
-    if not isinstance(run, dict):
-        return {"bucket": "missing"}
-    keys = (
-        "run_id",
-        "recorded_at",
-        "benchmark_id",
-        "case_id",
-        "run_group_id",
-        "arm_id",
-        "route",
-        "artifact_refs",
-        "status",
-        "score_status",
-        "official_score",
-        "official_passed",
-        "round_reward_count",
-        "round_success_observed",
-        "max_rounds_budget",
-        "app_server_goal_round_semantics",
-        "native_goal_session_policy",
-        "max_rounds_budget_applies_to",
-        "native_goal_initial_turn_budget",
-        "native_goal_same_thread_followup_budget",
-        "native_goal_independent_attempt_budget",
-        "native_goal_fresh_thread_per_independent_attempt",
-        "native_goal_official_reward_feedback_forwarded_to_worker",
-        "native_goal_verifier_output_forwarded_to_worker",
-        "official_feedback_blinded",
-        "reward_feedback_forwarded",
-        "failure_class",
-        "failure_scope",
-        "failure_labels",
-        "score_failure_attribution",
-        "failure_attribution_labels",
-        "attempt_lifecycle_phase",
-        "attempt_failure_label",
-        "attempt_failure_class",
-        "runner_return_status",
-        "runner_score_recovered_from_verifier_artifact",
-        "verifier_reward_artifact_recovery_status",
-        "verifier_reward_artifact_recovered",
-        "official_result_json_materialized",
-        "solution_quality_signals",
-        "task_setup_preflight",
-        "task_staging",
-        "repair_class",
-        "repair_priority",
-    )
-    summary = {key: run[key] for key in keys if key in run}
-    summary["bucket"] = _current_aggregate_bucket(run)
-    effective_failure_class = _current_aggregate_effective_failure_class(run)
-    if effective_failure_class:
-        summary["effective_failure_class"] = effective_failure_class
-    return summary
-
-
-def _current_aggregate_case_is_noncanonical_sanity_source(case: dict[str, Any]) -> bool:
-    runs = _active_ledger_runs(
-        [item for item in case.get("runs", []) if isinstance(item, dict)]
-    )
-    if not runs:
-        return False
-    saw_sanity_source_preflight = False
-    for run in runs:
-        if _ledger_score_value(run) is not None:
-            return False
-        if _current_aggregate_effective_failure_class(run) != (
-            "skillsbench_task_source_preflight_blocked"
-        ):
-            return False
-        preflight = run.get("task_setup_preflight")
-        if not isinstance(preflight, dict):
-            return False
-        if preflight.get("canonical_task_present") is not False:
-            return False
-        if _compact_text(preflight.get("status"), limit=120) != (
-            "task_missing_from_canonical_tasks"
-        ):
-            return False
-        if _compact_text(preflight.get("alternate_source_kind"), limit=120) != (
-            "experiments_sanity_tasks"
-        ):
-            return False
-        if preflight.get("alternate_source_supported_by_runner") is not False:
-            return False
-        saw_sanity_source_preflight = True
-    return saw_sanity_source_preflight
-
-
 def build_benchmark_run_ledger_current_aggregate(
     ledger: dict[str, Any],
     *,
@@ -3386,106 +3183,17 @@ def build_benchmark_run_ledger_current_aggregate(
     source_ledger_count: int = 1,
     exclude_noncanonical_sanity_sources: bool = True,
 ) -> dict[str, Any]:
-    """Build a current-case aggregate, preferring countable results over missing rows."""
-
-    normalized = _normalize_benchmark_run_ledger(dict(ledger))
-    benchmark = (
-        normalized.get("benchmarks", {}).get(benchmark_id)
-        if isinstance(normalized.get("benchmarks"), dict)
-        else None
+    from .benchmark_ledger_current import (
+        build_benchmark_run_ledger_current_aggregate as _build_current_aggregate,
     )
-    cases = benchmark.get("cases") if isinstance(benchmark, dict) else {}
-    if not isinstance(cases, dict):
-        cases = {}
-    if canonical_case_ids:
-        canonical_ids = [
-            _compact_text(case_id, limit=160)
-            for case_id in canonical_case_ids
-            if _compact_text(case_id, limit=160)
-        ]
-        if exclude_noncanonical_sanity_sources:
-            canonical_ids = [
-                case_id
-                for case_id in canonical_ids
-                if not (
-                    isinstance(cases.get(case_id), dict)
-                    and _current_aggregate_case_is_noncanonical_sanity_source(
-                        cases[case_id]
-                    )
-                )
-            ]
-    else:
-        canonical_ids = sorted(
-            case_id
-            for case_id, case in cases.items()
-            if not (
-                isinstance(case, dict)
-                and _current_aggregate_case_is_noncanonical_sanity_source(case)
-            )
-        )
-    distribution = {
-        "missing": 0,
-        "setup_runner_infra": 0,
-        "verifier_no_reward": 0,
-        "official_zero": 0,
-        "partial": 0,
-        "pass": 0,
-    }
-    cases_by_bucket = {bucket: [] for bucket in distribution}
-    case_best: dict[str, dict[str, Any]] = {}
-    deduped_run_ids: set[str] = set()
-    for case in cases.values():
-        if not isinstance(case, dict):
-            continue
-        for run in _active_ledger_runs(
-            [item for item in case.get("runs", []) if isinstance(item, dict)]
-        ):
-            run_id = _compact_text(run.get("run_id"), limit=80)
-            if run_id:
-                deduped_run_ids.add(run_id)
-    for case_id in canonical_ids:
-        case = cases.get(case_id)
-        runs = []
-        if isinstance(case, dict):
-            runs = _active_ledger_runs(
-                [item for item in case.get("runs", []) if isinstance(item, dict)]
-            )
-        best = max(runs, key=_current_aggregate_run_sort_key) if runs else None
-        summary = _current_aggregate_run_summary(best)
-        bucket = summary["bucket"]
-        distribution[bucket] = distribution.get(bucket, 0) + 1
-        cases_by_bucket.setdefault(bucket, []).append(case_id)
-        case_best[case_id] = summary
-    return {
-        "schema_version": BENCHMARK_RUN_LEDGER_CURRENT_AGGREGATE_SCHEMA_VERSION,
-        "benchmark_id": benchmark_id,
-        "canonical_total": len(canonical_ids),
-        "canonical_covered": sum(
-            count for bucket, count in distribution.items() if bucket != "missing"
-        ),
-        "distribution": distribution,
-        "cases_by_bucket": {
-            bucket: sorted(case_ids) for bucket, case_ids in cases_by_bucket.items()
-        },
-        "case_best": case_best,
-        "deduped_run_count": len(deduped_run_ids),
-        "source_ledger_files": source_ledger_count,
-        "selection_policy": {
-            "schema_version": "benchmark_run_ledger_current_selection_policy_v0",
-            "rule": "prefer_countable_official_result_over_missing_or_infra_rows",
-            "bucket_precedence": [
-                "pass",
-                "partial",
-                "official_zero",
-                "verifier_no_reward",
-                "setup_runner_infra",
-                "missing",
-            ],
-            "raw_logs_recorded": False,
-            "raw_task_text_recorded": False,
-            "source_paths_recorded": False,
-        },
-    }
+
+    return _build_current_aggregate(
+        ledger,
+        benchmark_id=benchmark_id,
+        canonical_case_ids=canonical_case_ids,
+        source_ledger_count=source_ledger_count,
+        exclude_noncanonical_sanity_sources=exclude_noncanonical_sanity_sources,
+    )
 
 
 def merge_benchmark_run_ledgers(
